requirements.txt
pyTelegramBotApi==4.5.1
python-dotenv==0.20.0
redis==4.3.1
